Refactor getProducts API into separate function
